Add built in qp_decode() for handling the quoted-printable encoding on incoming emails
The emaili::qp_decode() we were using wasn't actually fully functional,
and so a lot of =EA type strings were left in our output. There doesn't
appear to be a standard R library handling this, so we adapt a Stack
Overflow lookup table from Bob Rudis (@hrbrmstr), with thanks
https://stackoverflow.com/questions/40976213/decoding-quoted-printable-string-in-r

It's modified here to fix a double-decode bug when a =3D ('=') sequence
is followed by two characters that can be intepreted as a hex number
from \x40 upwards. See https://github.com/hrbrmstr/hrbrmisc/pull/6
for details.