Handle a set of input documents with mangled <html> tags
Within a ProQuest .eml there is a text/html segment, which itself contains individual embedded HTML documents for each article (normally up to 100).

Some of the article documents in my dataset don't have an <html> tag to start them - they have "html&gt;" instead. This breaks the parse, as the parser picks up the (correct) closing </html> tag for the embedded document and (because that document was never started) treats it as a closing tag for the *whole set of documents*. So every document after the mangled one is dropped.

We deal with this very specific grotty problem in an equally grotty way: we gsub() the broken character sequence with a fixed one before starting the parse. This is *unlikely* to break any other documents, but...